refactor: simplify profile image processing outputs
