+++ src/main/java/com/domisa/domisa_backend/profileimage/config/ProfileImageProcessingProperties.java
@@ -6,13 +6,17 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 public class ProfileImageProcessingProperties {
 
 	// 파생본 생성 규칙과 재시도 설정을 한곳에서 관리한다.
+	// 썸네일 정사각 크기다. 현재 300이면 300x300으로 중앙 크롭된다.
 	private int thumbnailSize = 300;
+	// 오리진 블러 이미지는 긴 변 기준으로 이 값까지 줄인다.
 	private int detailMaxSize = 1080;
+	// 썸네일 블러 강도다. 홀수 커널 크기이며 커질수록 더 흐려진다.
 	private int thumbnailBlurKernelSize = 13;
+	// 오리진 블러 강도다. 홀수 커널 크기이며 커질수록 더 흐려진다.
 	private int originBlurKernelSize = 17;
-	private double thumbnailJpegQuality = 0.85d;
-	private double originBlurJpegQuality = 0.9d;
+	// 한 번에 처리할 프로필 이미지 건수다.
 	private int batchSize = 10;
+	// 실패 시 최대 재시도 횟수다.
 	private int maxRetryCount = 5;
 
 	public int getThumbnailSize() {
@@ -45,22 +49,6 @@ public class ProfileImageProcessingProperties {
 
 	public void setOriginBlurKernelSize(int originBlurKernelSize) {
 		this.originBlurKernelSize = originBlurKernelSize;
-	}
-
-	public double getThumbnailJpegQuality() {
-		return thumbnailJpegQuality;
-	}
-
-	public void setThumbnailJpegQuality(double thumbnailJpegQuality) {
-		this.thumbnailJpegQuality = thumbnailJpegQuality;
-	}
-
-	public double getOriginBlurJpegQuality() {
-		return originBlurJpegQuality;
-	}
-
-	public void setOriginBlurJpegQuality(double originBlurJpegQuality) {
-		this.originBlurJpegQuality = originBlurJpegQuality;
 	}
 
 	public int getBatchSize() {
